Explicitly convert to numpy uint32 type where appropriate.
`dtype=int` converts to the C `long` data type. On 64b Linux, that defaults to
`int64_t`. On Windows however, even 64b, it is `int32_t`.

Using `dtype=int` on Windows will raise an OverflowError when trying to convert
uint32_t 0xFFFFFFFF values.

+++ python/fusion_engine_client/messages/solution.py
@@ -387,7 +387,7 @@ class GNSSInfoMessage(MessagePayload):
             'num_svs': np.array([int(m.num_svs) for m in messages], dtype=int),
             'corrections_age_sec': np.array([m.corrections_age_sec for m in messages]),
             'baseline_distance_m': np.array([m.baseline_distance_m for m in messages]),
-            'reference_station_id': np.array([int(m.reference_station_id) for m in messages], dtype=int),
+            'reference_station_id': np.array([int(m.reference_station_id) for m in messages], dtype=np.uint32),
             'gdop': np.array([m.gdop for m in messages]),
             'pdop': np.array([m.pdop for m in messages]),
             'hdop': np.array([m.hdop for m in messages]),
